Remove `fuubar` gem and remove custom rspec format setting
Instead of specifying this at the project level, we can stick to rspec
defaults. It's easy for individual developers to install custom
formatting gems and specify that format in their `.rspec-local` (which
can be excluded from git) if desired.

We get a slight bundle size reduction from removing the gem.