Simplify `Status.tagged_with_all` scope
Previously we were looping through all the supplied tag ids and
appending a new condition for each one so that the generated query was a
collection of AND'd EXISTS statements. Query times under this approach
are reasonable with small tag id counts, but grow as the tag id array
size grows.

The update uses a group by / having approach instead. Query times remain
fairly constant under this approach, with new IDs added to an `IN`
condition.