Extract admin/instances/delivery_* to namespace
Related changes all around the same general "make this more RESTful" theme.

For the endpoint which removes delivery errors, treat "the delivery errors" as a
restful concept and convert a "POST to clear_delivery_errors" to a "DELETE
errors#destroy" request.

For the endpoints which stop and restart deliveries (ie, hit the delivery
tracker), treat a "delivery pause" as a restful concept and move the "POST
stop_delivery" to "POST pauses#create" and the "POST restart_delivery" to
"DELETE pauses#destroy".